+++ website/frontend/src/components/blocks/Header/Header.scss
@@ -1,4 +1,4 @@
-@import '../../../assets/scss/variables';
+@import 'assets/scss/variables';
 
 .header {
   display: flex;
@@ -181,145 +181,6 @@
 
 .header__notifications {
   display: none;
-}
-
-.nodes {
-  position: fixed;
-  height: 100%;
-  top: 0;
-  left: 0;
-  background: linear-gradient(0deg, rgba(255, 255, 255, 0.12), rgba(255, 255, 255, 0.12)), #121212;
-  box-shadow: 0px 8px 10px rgba(0, 0, 0, 0.14), 0px 3px 14px rgba(0, 0, 0, 0.12), 0px 5px 5px rgba(0, 0, 0, 0.2);
-  padding: 25px 20px;
-  z-index: 100;
-  transition: all ease 300ms;
-  &.show {
-    left: 0;
-  }
-  .nodes__header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    margin-bottom: 30px;
-    .nodes__switch-button {
-      padding: 5px 15px 5px 20px;
-      display: flex;
-      flex-direction: row;
-      font-weight: bold;
-      align-items: center;
-      font-size: 16px;
-      line-height: 24px;
-      background: $successColor;
-      border-radius: 2px;
-      color: $textColor;
-      .nodes__switch-icon {
-        margin-right: 15px;
-      }
-    }
-  }
-
-  .nodes__wrap {
-    .nodes__item {
-      margin-bottom: 15px;
-      &:last-child {
-        margin-bottom: 0;
-      }
-      .nodes__item-caption {
-        font-weight: bold;
-        color: $textColor;
-        margin-bottom: 10px;
-        text-transform: uppercase;
-      }
-      .nodes__item-list {
-        .nodes__item-elem {
-          margin-bottom: 10px;
-          display: flex;
-          justify-content: space-between;
-          &:last-child {
-            margin-bottom: 0;
-          }
-          .nodes__item-icon {
-            width: 20px;
-            height: 20px;
-          }
-          .nodes__item-btn {
-            padding: 0;
-            display: flex;
-            align-items: center;
-          }
-          .nodes__item-text {
-            display: inline-block;
-            color: $textColor;
-            max-width: 230px;
-            overflow: hidden;
-            text-overflow: ellipsis;
-            white-space: nowrap;
-            text-align: left;
-            @media (max-width: 768px) {
-              max-width: 170px;
-            }
-          }
-          .nodes__item-remove {
-            padding: 0;
-            display: flex;
-            align-items: center;
-            img {
-              width: 20px;
-              height: 20px;
-            }
-          }
-          .nodes__item-choose {
-            margin-right: 10px;
-            .nodes__item-icon {
-              margin-right: 10px;
-            }
-          }
-          .nodes__item-btns {
-            display: flex;
-            align-items: center;
-            .nodes__item-btn {
-              margin-right: 10px;
-              &:last-child {
-                margin-right: 0;
-              }
-            }
-          }
-        }
-      }
-    }
-  }
-
-  .nodes__add {
-    margin-top: 20px;
-    display: flex;
-    flex-direction: column;
-    .nodes__add-input {
-      padding: 10px 20px;
-      border-radius: 2px;
-      border: none;
-      outline: none;
-      background-color: $mainBackground;
-      color: $textColorSecondary;
-      font-size: 16px;
-    }
-    .nodes__add-button {
-      margin-top: 10px;
-      border: none;
-      outline: none;
-      color: $textColor;
-      padding: 12px;
-      background: $successColor;
-      border-radius: 2px;
-      font-weight: bold;
-      font-size: 16px;
-      line-height: 24px;
-      color: $textColor;
-      &:disabled {
-        color: $textColor;
-        background: $bkgButtonColor;
-      }
-    }
-  }
 }
 
 @media (max-width: 768px) {
